requirements.txt
tensorflow>=2.1.0
tensorflow-hub>=0.7.0
bert-for-tf2